added loader on pages, navbar styling, fix footer
