+++ src/App.css
@@ -28,5 +28,13 @@
      text-decoration: none;
  }
 
- /* home */
+/* loader */
+.loader-prod {
+    height: 100vh;
+    font-size: 6rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    font-family: "Reggae One", cursive;
+  }
