+++ src/App.js
@@ -1,4 +1,4 @@
-import { NavBar } from './components/navBar'
+import { NavBar } from './components/NavBar/navBar'
 import { React } from 'react'
 import './App.css'
 import { Routes, Route } from "react-router-dom";
@@ -8,6 +8,7 @@ export default function App() {
   return (
     <div className="App">
       <NavBar />
+      <main style={{marginTop:"5rem", height:"100vh"}} >
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="products" element={<Products />} />
@@ -15,6 +16,7 @@ export default function App() {
         <Route path="wishlist" element={<WishList />} />
         <Route path="products/:productId" element={<ProductDetails />} />
       </Routes>
+      </main>
     </div>
   );
 }
