+++ src/components/Footer/footer.style.css
@@ -1,18 +1,23 @@
-footer{
-    height: 7rem;
-    background-color: black;
-    color: white;
-    display: flex;
-    justify-content: space-between;
+@import url("https://fonts.googleapis.com/css2?family=Reggae+One&display=swap");
+footer {
+  height: 7rem;
+  background-color: black;
+  color: white;
+  display: flex;
+  justify-content: space-between;
+  bottom: 0;
+  position: relative;
 }
-.container{
-    margin: 1rem;
-    padding: 1rem;
+.container {
+  margin: 1rem;
+  padding: 1rem;
 }
-p{
-    padding-bottom: 0.5rem;
+p {
+  padding-bottom: 0.5rem;
 }
-a{
-    text-decoration: none;
-    color: white;
+a {
+  text-decoration: none;
+  color: white;
 }
+
+
